+++ GLHelpers.h
@@ -84,8 +84,8 @@ void vao_bind(unsigned int vao_id);
 
 unsigned int vbo_create();
 void vbo_bind(unsigned int vbo_id, enum vbo_type type);
-void vbo_upload(int n_bytes, void *data, enum vbo_type type, enum vbo_hint hint);
-void vbo_reupload(int n_bytes, int offset, void *data, enum vbo_type type);
+void vbo_upload(size_t n_bytes, void *data, enum vbo_type type, enum vbo_hint hint);
+void vbo_reupload(size_t n_bytes, int offset, void *data, enum vbo_type type);
 
 
 // Programs:
